Changes to formatting and naming following code review.
