+++ src/main/webapp/electionInfo.html
@@ -19,6 +19,7 @@
 	  <link rel="stylesheet" href="style.css">
     <script src="auto-complete.js"></script>
     <script src="script.js"></script>
+    <script src="election-info.js"></script>
 
     </head>
