+++ src/main/webapp/index.html
@@ -14,6 +14,7 @@
 
     <link rel="stylesheet" href="style.css">
     <script src="script.js"></script>
+    <script src="welcome-page.js"></script>
 
   </head>
   <body onload="onWelcomePageLoad();">
