+++ src/main/webapp/welcome-page.js
@@ -1,0 +1,4 @@
+function onWelcomePageLoad(){
+  let links = [document.getElementById("home-link"), document.getElementById("elections-list-link")];
+  updateLinksWithQueryParams(links);
+}
