+++ src/test/java/com/google/sps/integration/ElectionInfoIT.java
@@ -1,6 +1,8 @@
 package com.google.sps.integration;
 
-import org.junit.*;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.ClassRule;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
