+++ src/test/java/com/google/sps/integration/WelcomePageIT.java
@@ -1,6 +1,10 @@
 package com.google.sps.integration;
 
-import org.junit.*;
+import org.junit.After;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.ClassRule;
+import org.junit.Test;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
